requirements.txt
httpx>=0.27.0
pyyaml>=6.0
aiofiles>=23.0
